move NeedlessReturn pass out to its own module and rename to ReturnPass
